refactor(features): split feature to two parts, merchant and customer
