+++ src/features/customer/components/BeautyCategoryWidget.tsx
