+++ src/features/customer/components/CombinedRatingComponent.tsx
