+++ src/features/customer/components/DetailedRatingsComponent.tsx
