+++ src/features/customer/components/ShoppingEntertainmentWidget.tsx
