+++ src/features/customer/components/StarRatingComponent.tsx
