+++ src/features/customer/components/index.ts
@@ -1,0 +1,22 @@
+// Home
+export { StudentPost } from './StudentPost';
+export { ActivityCard } from './ActivityCard';
+export { CityLocationButton } from './CityLocationButton';
+
+// Discover (formerly widgets)
+export { CategoryWidget } from './CategoryWidget';
+export { FoodCategoryWidget } from './FoodCategoryWidget';
+export { BeautyCategoryWidget } from './BeautyCategoryWidget';
+export { ShoppingEntertainmentWidget } from './ShoppingEntertainmentWidget';
+
+// Profile
+export { ProfileHeader } from './ProfileHeader';
+export { ProfileTabs } from './ProfileTabs';
+export { SavedPlaceCard } from './SavedPlaceCard';
+
+// Reviews
+export { default as CombinedRatingComponent } from './CombinedRatingComponent';
+export { default as DetailedRatingsComponent } from './DetailedRatingsComponent';
+export { default as ImageUploadGrid } from './ImageUploadGrid';
+export { default as ImageUploadWrapper } from './ImageUploadWrapper';
+export { default as StarRatingComponent } from './StarRatingComponent';
