+++ src/features/customer/constants/categories.ts
