+++ src/features/customer/constants/homeMockData.ts
