+++ src/features/customer/hooks/useAutoScroll.ts
