+++ src/features/customer/index.ts
@@ -1,0 +1,16 @@
+// Pages
+export { default as HomePage } from './pages/HomePage';
+export { default as DiscoverPage } from './pages/DiscoverPage';
+export { default as ProfilePage } from './pages/ProfilePage';
+export { default as PostPage } from './pages/PostPage';
+export { default as WriteReviewPage } from './pages/WriteReviewPage';
+
+// Layouts
+export { default as CustomerLayout } from './layout/CustomerLayout';
+export * from './layout/index'; // Export BottomNav, FAB, etc. from layout subfolder 
+
+// Contexts
+export * from './contexts/index';
+
+// Components (Optional: export often used components if needed)
+// export * from './components';
