+++ src/features/customer/types/index.ts
@@ -1,0 +1,4 @@
+export * from './home';
+export * from './discover';
+export * from './profile';
+export * from './reviews';
