+++ src/features/customer/utils/categoryUtils.ts
