+++ src/features/customer/utils/imageCompression.ts
