+++ src/features/customer/utils/uploadManager.ts
