--- src/features/discover/index.ts
@@ -1,8 +1,0 @@
-// Discover Feature exports
-export { default as DiscoverPage } from './pages/DiscoverPage';
-export * from './widgets';
-export * from './types';
-export * from './constants/categories';
-export * from './utils/categoryUtils';
-export * from './utils/recommendationUtils';
-export * from './hooks/useDebounce';
