--- src/features/discover/widgets/index.ts
@@ -1,5 +1,0 @@
-// Discover Widgets exports
-export { CategoryWidget } from './CategoryWidget';
-export { FoodCategoryWidget } from './FoodCategoryWidget';
-export { BeautyCategoryWidget } from './BeautyCategoryWidget';
-export { ShoppingEntertainmentWidget } from './ShoppingEntertainmentWidget';
