--- src/features/home/components/index.ts
@@ -1,4 +1,0 @@
-// Home Components exports
-export { StudentPost } from './StudentPost';
-export { ActivityCard } from './ActivityCard';
-export { CityLocationButton } from './CityLocationButton';
