--- src/features/home/index.ts
@@ -1,6 +1,0 @@
-// Home Feature exports
-export { default as HomePage } from './pages/HomePage';
-export * from './components';
-export * from './types';
-export * from './utils/postUtils';
-export * from './hooks/useAutoScroll';
