--- src/features/profile/components/index.ts
@@ -1,4 +1,0 @@
-// Profile Components exports
-export { ProfileHeader } from './ProfileHeader';
-export { ProfileTabs } from './ProfileTabs';
-export { SavedPlaceCard } from './SavedPlaceCard';
