--- src/features/profile/index.ts
@@ -1,4 +1,0 @@
-// Profile Feature exports
-export { default as ProfilePage } from './pages/ProfilePage';
-export * from './components';
-export * from './types';
