--- src/features/reviews/components/index.ts
@@ -1,10 +1,0 @@
-// Component exports
-export { default as StarRatingComponent } from './StarRatingComponent';
-export type { StarRatingProps } from './StarRatingComponent';
-export { default as DetailedRatingsComponent } from './DetailedRatingsComponent';
-export type { DetailedRatingsProps } from './DetailedRatingsComponent';
-export { default as CombinedRatingComponent } from './CombinedRatingComponent';
-export type { CombinedRatingProps } from './CombinedRatingComponent';
-export { default as ImageUploadGrid } from './ImageUploadGrid';
-export type { ImageUploadGridProps } from './ImageUploadGrid';
-export { default as ImageUploadWrapper } from './ImageUploadWrapper';
