--- src/features/reviews/index.ts
@@ -1,6 +1,0 @@
-// Reviews Feature exports
-export { default as PostPage } from './pages/PostPage';
-export { default as WriteReviewPage } from './pages/WriteReviewPage';
-export * from './types';
-export * from './utils/reviewUtils';
-export * from './contexts';
